Refactor functions and write collate function
